tlhclk1312@gmail.com 123qweasd
Update 16-07-2017

+++ account_panel/views.py
@@ -2,45 +2,59 @@
 from __future__ import unicode_literals
 from .models import TransactionInfoModel,PersonAssetInfoModel,AccountInfoModel,BillInfoModel
 from .forms import TransactionInfoForm,PersonAssetInfoForm,AccountInfoForm,BillInfoForm
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect,HttpResponse
+import datetime
+import re
 
 
 def options_menu(request):
     return render(request,'account_panel/option_menu.html')
 
 def add_transaction(request):
-    formtransaction=TransactionInfoForm()
-    if request.method=='POST':
-        formtransaction=TransactionInfoForm(request.POST)
-        if formtransaction.is_valid():
-            account=formtransaction.cleaned_data.get('account_no')
-            amount=formtransaction.cleaned_data.get('transaction_amount')
-            d_c=formtransaction.cleaned_data.get('transaction_type')
-            transaction_accountsync(account,amount,d_c)
-            formtransaction.save()
-            return redirect('../')
-    return render(request,'account_panel/default_form.html',{'form':formtransaction})
+    if request.user.has_perm('account_panel.add_transactioninfomodel'):
+        formtransaction=TransactionInfoForm()
+        if request.method=='POST':
+            formtransaction=TransactionInfoForm(request.POST)
+            if formtransaction.is_valid():
+                account=formtransaction.cleaned_data.get('account_no')
+                amount=formtransaction.cleaned_data.get('transaction_amount')
+                transaction_type=formtransaction.cleaned_data.get('transaction_type')
+                transaction_desc=formtransaction.cleaned_data.get('transaction_desc')
+                transaction_accountsync(account,amount,transaction_type,transaction_desc)
+                formtransaction.save()
+                return redirect('../')
+        return render(request,'account_panel/default_form.html',{'form':formtransaction})
+    else: return HttpResponse('You has No authorizations')
 
 def detail_transaction(request,transaction_no):
-    transaction=TransactionInfoModel.objects.get(pk=transaction_no)
-    return render(request,'account_panel/detail_transaction.html',{'transaction':transaction})
+    if request.user.has_perm('account_panel.add_transactioninfomodel'):
+        transaction=TransactionInfoModel.objects.get(pk=transaction_no)
+        all_fields=TransactionInfoModel._meta.get_fields()
+        return render(request,'account_panel/detail_transaction.html',{'transaction':transaction})
+    else: return HttpResponse('You has No authorizations')
 
 def table_transaction(request):
-    transaction_list=TransactionInfoModel.objects.all()
-    return render(request,'account_panel/table_transaction.html',{'transaction_list':transaction_list})
+    if request.user.has_perm('account_panel.view_transactioninfomodel'):
+        transaction_list=TransactionInfoModel.objects.all()
+        return render(request,'account_panel/table_transaction.html',{'transaction_list':transaction_list})
+    else: return HttpResponse('You has No authorizations')
 
 def edit_transaction(request,transaction_no):
-    formtransaction=TransactionInfoForm(instance=TransactionInfoModel.objects.get(pk=transaction_no))
-    if request.method=='POST':
-        formtransaction=TransactionInfoForm(request.POST,instance=TransactionInfoModel.objects.get(pk=transaction_no))
-        if formtransaction.is_valid():
-            formtransaction.save()
-            return redirect('../')
-    return render(request,'account_panel/default_form.html',{'form':formtransaction})
+    if request.user.has_perm('account_panel.add_transactioninfomodel'):
+        formtransaction=TransactionInfoForm(instance=TransactionInfoModel.objects.get(pk=transaction_no))
+        if request.method=='POST':
+            formtransaction=TransactionInfoForm(request.POST,instance=TransactionInfoModel.objects.get(pk=transaction_no))
+            if formtransaction.is_valid():
+                formtransaction.save()
+                return redirect('../')
+        return render(request,'account_panel/default_form.html',{'form':formtransaction})
+    else: return HttpResponse('You has No authorizations')
 
 def delete_transaction(request,transaction_no):
-    TransactionInfoModel.objects.get(pk=transaction_no).delete()
-    return redirect('../')
+    if request.user.has_perm('account_panel.add_transactioninfomodel'):
+        TransactionInfoModel.objects.get(pk=transaction_no).delete()
+        return redirect('../')
+    else: return HttpResponse('You has No authorizations')
 
 def add_asset(request):
     formasset=PersonAssetInfoForm()
@@ -95,7 +109,7 @@ def edit_account(request,account_no):
         formaccount=AccountInfoForm(request.POST,instance=AccountInfoModel.objects.get(pk=account_no))
         if formaccount.is_valid():
             formaccount.save()
-            return redirect('../')
+            return redirect('../../')
     return render(request,'account_panel/default_form.html',{'form':formaccount})
 
 def delete_account(request,account_no):
@@ -107,6 +121,7 @@ def add_bill(request):
     if request.method=='POST':
         formbill=BillInfoForm(request.POST)
         if formbill.is_valid():
+            bill_transactionsync(formbill)
             formbill.save()
             return redirect('../')
     return render(request,'account_panel/default_form.html',{'form':formbill})
@@ -132,11 +147,31 @@ def delete_bill(request,bill_no):
     BillInfoModel.objects.get(pk=bill_no).delete()
     return redirect('../')
 
-def transaction_accountsync(account,amount,d_c):
-    if d_c=='1' or d_c=='7':
+def transaction_accountsync(account_no,amount,transaction_type,transaction_desc):
+    if transaction_type=='1' or transaction_type=='7':
         x=1
+    elif transaction_type=='8':
+        if re.search('1701\w{3}',transaction_desc):
+            print (re.search('1701\w{3}',transaction_desc).group())
+            transaction_to = re.search('1701\w{3}',transaction_desc).group()
+            x = 1
+            asset = PersonAssetInfoModel.objects.filter(person_id=transaction_to)[0]
+            asset.asset_amount = str(float(asset.asset_amount) + x * float(amount))
+            asset.asset_debt = str(float(asset.asset_debt) - x * float(amount))
+            asset.save()
+        else:
+            return HttpResponse('Error','Transaction Description has some Errors')
     else:
         x=-1
-    account.account_amount=str(float(account.account_amount)+x*float(amount))
-    account.save()
+    account_no.account_amount=str(float(account_no.account_amount)+x*float(amount))
+    account_no.save()
     return redirect('../')
+
+def bill_transactionsync(formbill):
+    account_no=AccountInfoModel.objects.filter(account_name='Nakit')[0]
+    amount=formbill.cleaned_data.get('bill_amount')
+    description=formbill.cleaned_data.get('bill_desc')
+    transaction=TransactionInfoModel(account_no=account_no,transaction_type='6',transaction_amount=amount,transaction_time=datetime.datetime.now(),transaction_desc=description)
+    transaction_accountsync(account_no,amount,transaction_type='6',transaction_desc=description)
+    transaction.save()
+
